LPS-203260 SF Extract the interfaces, no logic changes
